Authenticate controller made, admin specs fixed
