+++ app/controllers/admin_controller.rb
@@ -1,7 +1,7 @@
 class AdminController < ApplicationController
   include AdminHelper
+
   before_action :admin_authenticate
-  skip_before_action :admin_authenticate, only: :login
 
   def admin_authenticate
     unless logged_in?
@@ -11,15 +11,6 @@ class AdminController < ApplicationController
 
   def index
     render :index
-  end
-
-  # Get request
-  def login_form
-    render :login
-  end
-
-  # Post request
-  def login
   end
 
 end
