+++ app/helpers/authenticate_helper.rb
@@ -1,0 +1,2 @@
+module AuthenticateHelper
+end
